+++ src/electron/frontend/core/components/JSONSchemaInput.js
@@ -3,17 +3,17 @@ import { unsafeHTML } from "lit/directives/unsafe-html.js";
 import { FilesystemSelector } from "./FileSystemSelector";
 
 import { BasicTable } from "./BasicTable";
-import { header, tempPropertyKey, tempPropertyValueKey } from "./forms/utils";
+import { header, tempPropertyKey, tempPropertyValueKey } from "../../utils/forms";
 
 import { Button } from "./Button";
 import { List } from "./List";
 import { Modal } from "./Modal";
 
-import { capitalize } from "./forms/utils";
+import { capitalize } from "../../utils/forms";
 import { JSONSchemaForm, getIgnore } from "./JSONSchemaForm";
 import { Search } from "./Search";
 import tippy from "tippy.js";
-import { merge } from "./pages/utils";
+import { merge } from "../../utils/pages";
 import { OptionalSection } from "./OptionalSection";
 import { InspectorListItem } from "./preview/inspector/InspectorList";
 import { renderDateTime, resolveDateTime } from "./DateTimeSelector";
