+++ src/electron/frontend/core/components/pages/guided-mode/setup/GuidedNewDatasetInfo.js
@@ -6,9 +6,10 @@ import { validateOnChange } from "../../../../validation/index.js";
 
 import projectGeneralSchema from "../../../../../../../schemas/json/project/general.json" assert { type: "json" };
 import projectGlobalSchema from "../../../../../../../schemas/json/project/globals.json" assert { type: "json" };
-import { merge } from "../../utils";
 import { onThrow } from "../../../../errors";
-import { header } from "../../../forms/utils";
+
+import { merge } from "../../../../../utils/pages";
+import { header } from "../../../../../utils/forms";
 
 const projectMetadataSchema = merge(projectGlobalSchema, projectGeneralSchema);
