+++ src/electron/frontend/core/components/pages/guided-mode/setup/Preform.js
@@ -2,7 +2,7 @@ import { html } from "lit";
 import { JSONSchemaForm } from "../../../JSONSchemaForm.js";
 import { Page } from "../../Page.js";
 import { onThrow } from "../../../../errors";
-import { merge } from "../../utils.js";
+import { merge } from "../../../../../utils/pages.js";
 import timezoneSchema from "../../../../../../../schemas/timezone.schema";
 
 // ------------------------------------------------------------------------------
