+++ src/electron/frontend/core/components/pages/inspect/InspectPage.js
@@ -3,12 +3,13 @@ import { Page } from "../Page.js";
 import { onThrow } from "../../../errors";
 import { Button } from "../../Button.js";
 
-import { run } from "../guided-mode/options/utils.js";
+import { run } from "../../../../utils/requests.js";
 import { Modal } from "../../Modal";
 import { getSharedPath, truncateFilePaths } from "../../preview/NWBFilePreview.js";
 import { InspectorList, InspectorLegend } from "../../preview/inspector/InspectorList.js";
-import { download } from "./utils";
-import { createProgressPopup } from "../../utils/progress.js";
+import { download } from "../../../../utils/download.js";
+
+import { createProgressPopup } from "../../../../utils/progress.js";
 
 import { ready } from "../../../../../../schemas/dandi-upload.schema";
 import { JSONSchemaForm } from "../../JSONSchemaForm.js";
