+++ src/electron/frontend/core/components/preview/NWBFilePreview.js
@@ -2,7 +2,8 @@ import { LitElement, css, html } from "lit";
 import { InspectorList } from "./inspector/InspectorList";
 import { Neurosift, getURLFromFilePath } from "./Neurosift";
 import { unsafeHTML } from "lit/directives/unsafe-html.js";
-import { run } from "../pages/guided-mode/options/utils";
+import { run } from "../../../utils/requests.js";
+
 import { until } from "lit/directives/until.js";
 import { InstanceManager } from "../InstanceManager";
 import { path } from "../../../utils/electron.js";
