+++ src/electron/frontend/core/components/table/cells/input.ts
@@ -6,7 +6,7 @@ import { Modal } from "../../Modal.js";
 import { SimpleTable } from "../../SimpleTable.js";
 import { JSONSchemaInput } from "../../JSONSchemaInput.js";
 
-import { header } from "../../forms/utils.js";
+import { header } from "../../../../utils/forms.js";
 
 export class NestedEditor extends LitElement {
