--- src/electron/frontend/core/components/utils/useGlobalStyles.js
@@ -1,16 +1,0 @@
-import { css } from "lit";
-
-const useGlobalStyles = (componentCSS, condition, toApply = true) => {
-    if (!toApply || !condition) return css([componentCSS]);
-
-    const sheets = Object.values(document.styleSheets);
-    const selectedSheets = condition instanceof Function ? sheets.filter(condition) : sheets;
-    const rules = selectedSheets.map((sheet) =>
-        Object.values(sheet.cssRules)
-            .map((rule) => rule.cssText)
-            .join("\n")
-    );
-    return css([componentCSS, ...rules]);
-};
-
-export default useGlobalStyles;
