+++ src/electron/frontend/core/validation/backend-configuration.ts
@@ -1,4 +1,4 @@
-import { humanReadableBytes } from "../components/utils/size";
+import { humanReadableBytes } from "../../utils/size";
 
 const prod = (arr: number[]) => arr.reduce((accumulator, currentValue) => accumulator * currentValue, 1);
