+++ src/electron/frontend/utils/size.ts
