+++ src/electron/frontend/utils/tables.ts
